crawler/requirements.txt
# YouTube 댓글 크롤링 및 감성분석
google-api-python-client==2.108.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
transformers==4.36.2
torch==2.1.2
opensearch-py==2.4.2

# 벡터 임베딩 및 유사도 계산
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3

# 텍스트 처리
konlpy==0.6.0
nltk==3.8.1

# 유틸리티
tqdm==4.66.1